+++ src/conponents/DisplayText.tsx
@@ -1,4 +1,3 @@
-// Detailsっていう型つくったからImportしてpropsにあてて
 import { Details } from '@/types/Details'
 import styles from '@/styles/DisplayText.module.css'
 import Link from 'next/link'
@@ -12,7 +11,7 @@ export default function DisplayText(props: Details) {
         <p>{props.title}</p>
       </div>
       <div className={styles.DisplayText_detail}>
-      <Link href={props.url || '/'}>{props.detail}</Link>
+        <Link href={props.url || '/'}>{props.detail}</Link>
       </div>
     </div>
   )
