+++ src/conponents/NewCard.tsx
@@ -9,19 +9,19 @@ import { Cards } from '@/types/Cards';
 import styles from '@/styles/Cards.module.css'
 import DisplayCard from './Card';
 
-export default function MakeNewCard(){
-    return(
+export default function MakeNewCard() {
+    return (
         <DisplayCard
-        id={""}
-        name={"＋新規作成"}
-        organization={""}
-        x={""}
-        instagram={""}
-        others={""}
-        urlEnabled={false}
-        textColor= {"#969696"}
-        bgColor={"#FFF"}
-        onClickHandler={()=>{}}
+            id=""
+            name="＋新規作成"
+            organization=""
+            x=""
+            instagram=""
+            others=""
+            urlEnabled={false}
+            textColor="#969696"
+            bgColor="#FFF"
+            onClickHandler={() => { }}
         />
     )
 }
